support smallcheck-1.1.3
smallcheck-1.1.3 defines instances for Word, so we have to remove them
in order to avoid a compilation error. As a result, smallcheck-series
is no longer compatible with versions of smallcheck order than 1.1.3,
because if we try to compile the new Word-less version of
smallcheck-series with an older smallcheck, we won't have an instance
for Word.